// web/src/shared/api/generated/model/questionDecisionResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface QuestionDecisionResult {
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  version: number | string;
  outcome: string;
  /** @nullable */
  reason: string | null;
  difficulty: string;
  /** @nullable */
  difficultyChangedFrom: string | null;
  decidedBy: string;
  /** @nullable */
  decidedByName: string | null;
  decidedAt: string;
}